Add portable criu test for ubuntu and ubi - Add portable criu test with keep checkpoint and restore functions - Add images for both ubuntu and ubi - These tests will only run when jdk supports criu

+++ external/dockerfile_functions.sh
@@ -89,8 +89,14 @@ print_image_args() {
         tag="${version}-jdk"
     fi
     if [[ "${vm}" == "openj9" ]]; then
-        image_name="ibm-semeru-runtimes"
-        tag=open-${tag}
+        if  [[ "${os}" == "ubuntu" ]]; then
+            image_name="ibm-semeru-runtimes"
+            tag=open-${tag}
+        else
+            # os is ubi
+            image_name="registry.access.redhat.com/ubi8/ubi"
+            tag="8.5"
+        fi
     fi
     image="${image_name}:${tag}"
 
@@ -127,27 +133,88 @@ print_ubuntu_pkg() {
             "\n" >> ${file}
 }
 
+# Select the ubuntu OS packages
+print_ubi_pkg() {
+    local file=$1
+    local packages=$2
+
+    echo -e "RUN dnf install -y ${packages} \\" \
+            "\n\t&& dnf clean all " >> ${file}
+    echo -e "\nENV LANG='en_US.UTF-8' LANGUAGE='en_US:en' LC_ALL='en_US.UTF-8'" >> ${file}
+}
+
+# Install JDK
+print_jdk_install() {
+    local file=$1
+    local os=$2
+    local platform=$3
+
+    # JDK install is only needed for ubi image, and the installed jdk requires criu support, and will be replaced by mounted jdk during test
+    echo -e "\nRUN set -eux; \\" \
+            "\n\t case \"${platform}\" in \\" \
+            "\n\t  *x86-64*) \\" \
+            "\n\t     BUILD_ID_LAST_SUCCESS=\$(wget -qO- https://openj9-jenkins.osuosl.org/job/Build_JDK11_x86-64_linux_criu_Nightly/lastSuccessfulBuild/buildNumber); \\" \
+            "\n\t     BINARY_URL=\$(wget -qO- https://openj9-jenkins.osuosl.org/job/Build_JDK11_x86-64_linux_criu_Nightly/lastSuccessfulBuild/consoleText | grep -Po \"(?<=Deploying artifact: )https://openj9-artifactory.osuosl.org/artifactory/ci-openj9/Build_JDK11_x86-64_linux_criu_Nightly/\${BUILD_ID_LAST_SUCCESS}/OpenJ9-JDK11-x86-64_linux_criu.*tar.gz\"); \\" \
+            "\n\t     ;; \\" \
+            "\n\t  *) \\" \
+            "\n\t     echo \"Unsupported platform \"; \\" \
+            "\n\t     exit 1; \\" \
+            "\n\t     ;; \\" \
+            "\n\t esac; \\" \
+            "\n\t curl -LfsSo /tmp/openjdk.tar.xz \${BINARY_URL}; \\" \
+            "\n\t mkdir /tmp/jdk-extract; \\" \
+            "\n\t cd /tmp/jdk-extract; \\" \
+            "\n\t tar -xf /tmp/openjdk.tar.xz --strip-components=1; \\" \
+            "\n\t ./bin/jlink --no-header-files --no-man-pages --compress=2 --add-modules java.base,java.compiler,java.datatransfer,java.desktop,java.instrument,java.logging,java.management,java.management.rmi,java.naming,java.net.http,java.prefs,java.rmi,java.scripting,java.se,java.security.jgss,java.security.sasl,java.smartcardio,java.sql,java.sql.rowset,java.transaction.xa,java.xml,java.xml.crypto,jdk.accessibility,jdk.attach,jdk.charsets,jdk.compiler,jdk.crypto.cryptoki,jdk.crypto.ec,jdk.dynalink,jdk.editpad,jdk.httpserver,jdk.internal.ed,jdk.internal.jvmstat,jdk.internal.le,jdk.internal.opt,jdk.jartool,jdk.javadoc,jdk.jcmd,jdk.jconsole,jdk.jdeps,jdk.jdi,jdk.jdwp.agent,jdk.jlink,jdk.jshell,jdk.jsobject,jdk.localedata,jdk.management,jdk.management.agent,jdk.naming.dns,jdk.naming.ldap,jdk.naming.rmi,jdk.net,jdk.pack,jdk.rmic,jdk.scripting.nashorn,jdk.scripting.nashorn.shell,jdk.sctp,jdk.security.auth,jdk.security.jgss,jdk.unsupported,jdk.unsupported.desktop,jdk.xml.dom,jdk.zipfs,openj9.criu,openj9.cuda,openj9.dataaccess,openj9.dtfj,openj9.dtfjview,openj9.gpu,openj9.jvm,openj9.sharedclasses,openj9.traceformat,openj9.zosconditionhandling --output /opt/java/openjdk; \\" \
+            "\n\t rm -rf /tmp/jdk-extract; \\" \
+            "\n\t rm -rf /tmp/openjdk.tar.xz; " \
+             "\n" >> ${file}
+
+    echo -e "\nENV JAVA_HOME=/opt/java/openjdk \\" \
+            "\n\t PATH=\"/opt/java/openjdk/bin:\$PATH\" " \
+            "\n" >> ${file}
+
+    echo -e "\nENV JAVA_TOOL_OPTIONS=\"-XX:+IgnoreUnrecognizedVMOptions -XX:+IdleTuningGcOnIdle\" " \
+            "\n" >> ${file}
+
+    echo -e "\nENV RANDFILE=/tmp/.rnd  \\" \
+            "\n\t OPENJ9_JAVA_OPTIONS=\"-XX:+IgnoreUnrecognizedVMOptions -XX:+IdleTuningGcOnIdle -Dosgi.checkConfiguration=false\" " \
+            "\n" >> ${file}
+
+}
+
 # Install Ant
 print_ant_install() {
     local file=$1
     local ant_version=$2
+    local os=$3
 
     echo -e "ARG ANT_VERSION=${ant_version}" \
           "\nENV ANT_VERSION=\$ANT_VERSION" \
           "\nENV ANT_HOME=/opt/ant" \
-          "\n\n# Install Ant" \
-          "\nRUN wget --no-verbose --no-check-certificate --no-cookies http://archive.apache.org/dist/ant/binaries/apache-ant-\${ANT_VERSION}-bin.tar.gz \\" \
-          "\n\t&& wget --no-verbose --no-check-certificate --no-cookies http://archive.apache.org/dist/ant/binaries/apache-ant-\${ANT_VERSION}-bin.tar.gz.sha512 \\" >> ${file}
+          "\n\n# Install Ant" >> ${file}
 
-    echo -e "\t&& echo \"\$(cat apache-ant-\${ANT_VERSION}-bin.tar.gz.sha512) apache-ant-\${ANT_VERSION}-bin.tar.gz\" | sha512sum -c \\" >> ${file}
-
-    echo -e "\t&& tar -zvxf apache-ant-\${ANT_VERSION}-bin.tar.gz -C /opt/ \\" \
+    if [[ "${os}" == *"ubi"* ]]; then
+        echo -e "\nRUN cd /tmp \\" \
+            "\n\t&& wget --progress=dot:mega -O ant.zip https://archive.apache.org/dist/ant/binaries/apache-ant-\${ANT_VERSION}-bin.zip \\" \
+            "\n\t&& unzip -q ant.zip -d /opt \\" \
             "\n\t&& ln -s /opt/apache-ant-\${ANT_VERSION} /opt/ant \\" \
-            "\n\t&& rm -f apache-ant-\${ANT_VERSION}-bin.tar.gz \\" \
-            "\n\t&& rm -f apache-ant-\${ANT_VERSION}-bin.tar.gz.sha512" \
-            "\n\n# Add Ant to PATH" \
-            "\nENV PATH \${PATH}:\${ANT_HOME}/bin" \
+            "\n\t&& ln -s /opt/ant/bin/ant /usr/bin/ant " \
             "\n" >> ${file}
+    else
+        echo -e "\nRUN wget --no-verbose --no-check-certificate --no-cookies http://archive.apache.org/dist/ant/binaries/apache-ant-\${ANT_VERSION}-bin.tar.gz \\" \
+            "\n\t&& wget --no-verbose --no-check-certificate --no-cookies http://archive.apache.org/dist/ant/binaries/apache-ant-\${ANT_VERSION}-bin.tar.gz.sha512 \\" >> ${file}
+
+        echo -e "\t&& echo \"\$(cat apache-ant-\${ANT_VERSION}-bin.tar.gz.sha512) apache-ant-\${ANT_VERSION}-bin.tar.gz\" | sha512sum -c \\" >> ${file}
+
+        echo -e "\t&& tar -zvxf apache-ant-\${ANT_VERSION}-bin.tar.gz -C /opt/ \\" \
+                "\n\t&& ln -s /opt/apache-ant-\${ANT_VERSION} /opt/ant \\" \
+                "\n\t&& rm -f apache-ant-\${ANT_VERSION}-bin.tar.gz \\" \
+                "\n\t&& rm -f apache-ant-\${ANT_VERSION}-bin.tar.gz.sha512" \
+                "\n\n# Add Ant to PATH" \
+                "\nENV PATH \${PATH}:\${ANT_HOME}/bin" \
+                "\n" >> ${file}
+    fi
 }
 
 # Install Ant Contrib
@@ -156,10 +223,10 @@ print_ant_contrib_install() {
     local ant_contrib_version=$2
 
     echo -e "ARG ANT_CONTRIB_VERSION=${ant_contrib_version}" \
-          "\nENV ANT_CONTRIB_VERSION=\$ANT_CONTRIB_VERSION" \
-          "\n\n# Install Ant Contrib" \
-          "\nRUN wget --no-verbose --no-check-certificate --no-cookies https://sourceforge.net/projects/ant-contrib/files/ant-contrib/\${ANT_CONTRIB_VERSION}/ant-contrib-\${ANT_CONTRIB_VERSION}-bin.tar.gz \\" \
-          "\n\t&& wget --no-verbose --no-check-certificate --no-cookies https://sourceforge.net/projects/ant-contrib/files/ant-contrib/\${ANT_CONTRIB_VERSION}/ant-contrib-\${ANT_CONTRIB_VERSION}-bin.tar.gz.md5 \\" >> ${file}
+        "\nENV ANT_CONTRIB_VERSION=\$ANT_CONTRIB_VERSION" \
+        "\n\n# Install Ant Contrib" \
+        "\nRUN wget --no-verbose --no-check-certificate --no-cookies https://sourceforge.net/projects/ant-contrib/files/ant-contrib/\${ANT_CONTRIB_VERSION}/ant-contrib-\${ANT_CONTRIB_VERSION}-bin.tar.gz \\" \
+        "\n\t&& wget --no-verbose --no-check-certificate --no-cookies https://sourceforge.net/projects/ant-contrib/files/ant-contrib/\${ANT_CONTRIB_VERSION}/ant-contrib-\${ANT_CONTRIB_VERSION}-bin.tar.gz.md5 \\" >> ${file}
 
     echo -e "\t&& echo \"\$(cat ant-contrib-\${ANT_CONTRIB_VERSION}-bin.tar.gz.md5) ant-contrib-\${ANT_CONTRIB_VERSION}-bin.tar.gz\" | md5sum -c \\" >> ${file}
 
@@ -280,23 +347,43 @@ print_python_install() {
 print_criu_install() {
     local file=$1
     local criu_version=$2
+    local os=$3
+    local platform=$4
 
-    echo -e "ARG CRIU_VERSION=${criu_version}" \
-          "\nENV CRIU_VERSION=\$CRIU_VERSION" \
-          "\n\n# Install criu and set capabilities" \
-          "\nRUN wget --progress=dot:mega -O v\${CRIU_VERSION}.tar.gz https://github.com/checkpoint-restore/criu/archive/refs/tags/v\${CRIU_VERSION}.tar.gz \\" >> ${file}
-    
-    echo -e "\t&& tar -xzf v\${CRIU_VERSION}.tar.gz \\" \
-            "\n\t&& cd criu-\${CRIU_VERSION}  \\" \
-            "\n\t&& make \\" \
-            "\n\t&& make install \\" \
-            "\n\t&& cd .. \\" \
-            "\n\t&& rm -rf v\${CRIU_VERSION}.tar.gz criu-\${CRIU_VERSION} \\" \
-            "\n\t&& setcap cap_chown,cap_dac_override,cap_dac_read_search,cap_fowner,cap_fsetid,cap_kill,cap_setgid,cap_setuid,cap_setpcap,cap_net_admin,cap_sys_chroot,cap_sys_ptrace,cap_sys_admin,cap_sys_resource,cap_sys_time,cap_audit_control=eip /usr/local/sbin/criu" \
-            "\n\n# Run ldconfig to discover newly installed shared libraries" \
-            "\nRUN for dir in lib lib64 ; do echo /usr/local/$dir ; done > /etc/ld.so.conf.d/usr-local.conf \\" \
-            "\n\t&& ldconfig " \
-            "\n" >> ${file}
+    if [[ "${os}" == *"ubi"* ]]; then
+        if [[ "${platform}" == *"x86-64"* ]]; then
+            echo -e "\nRUN wget -O /usr/sbin/criu https://public.dhe.ibm.com/ibmdl/export/pub/software/openliberty/runtime/criu-build/b1/criu \\" \
+                    "\n\t&& wget -O /usr/lib64/libcriu.so.2.0 https://public.dhe.ibm.com/ibmdl/export/pub/software/openliberty/runtime/criu-build/b1/libcriu.so " \
+                    "\n" >> ${file}
+
+            echo -e "\nRUN chmod a+x /usr/sbin/criu \\" \
+                    "\n\t&& setcap cap_checkpoint_restore,cap_net_admin,cap_sys_ptrace=eip /usr/sbin/criu \\" \
+                    "\n\t&& cd /usr/lib64 \\" \
+                    "\n\t&& ln -s libcriu.so.2.0 libcriu.so \\" \
+                    "\n\t&& ln -s libcriu.so.2.0 libcriu.so.2 \\" \
+                    "\n\t&& cd / "  >> ${file}
+        else
+            echo "Criu binaries are not available for Platform ${platform}!"
+            exit 1
+        fi
+    else # for ubuntu
+        echo -e "ARG CRIU_VERSION=${criu_version}" \
+                "\nENV CRIU_VERSION=\$CRIU_VERSION" \
+                "\n\n# Install criu and set capabilities" \
+                "\nRUN wget --progress=dot:mega -O v\${CRIU_VERSION}.tar.gz https://github.com/checkpoint-restore/criu/archive/refs/tags/v\${CRIU_VERSION}.tar.gz \\" >> ${file}
+        
+        echo -e "\t&& tar -xzf v\${CRIU_VERSION}.tar.gz \\" \
+                "\n\t&& cd criu-\${CRIU_VERSION}  \\" \
+                "\n\t&& make \\" \
+                "\n\t&& make install \\" \
+                "\n\t&& cd .. \\" \
+                "\n\t&& rm -rf v\${CRIU_VERSION}.tar.gz criu-\${CRIU_VERSION} \\" \
+                "\n\t&& setcap cap_chown,cap_dac_override,cap_dac_read_search,cap_fowner,cap_fsetid,cap_kill,cap_setgid,cap_setuid,cap_setpcap,cap_net_admin,cap_sys_chroot,cap_sys_ptrace,cap_sys_admin,cap_sys_resource,cap_sys_time,cap_audit_control=eip /usr/local/sbin/criu" \
+                "\n\n# Run ldconfig to discover newly installed shared libraries" \
+                "\nRUN for dir in lib lib64 ; do echo /usr/local/$dir ; done > /etc/ld.so.conf.d/usr-local.conf \\" \
+                "\n\t&& ldconfig " \
+                "\n" >> ${file}
+    fi
 }
 
 
@@ -359,6 +446,10 @@ print_test_files() {
                 "\nCOPY ${test}/test.sh /test.sh" \
                 "\nCOPY test_base_functions.sh test_base_functions.sh\n" >> ${file}
     fi
+    if [[ "${test}" == *"portable"* ]]; then
+        echo -e "# This is the script used to restore portable test later." \
+            "\nCOPY ${test}/test_restore.sh /test_restore.sh" >> ${file}
+    fi
     if [[ ! -z ${localPropertyFile} ]]; then
         echo -e "# This local property file is needed to set up user preferred properties." \
             "\nCOPY ${test}/${localPropertyFile} \${TEST_HOME}/${localPropertyFile}\n" >> ${file}
@@ -386,6 +477,10 @@ print_clone_project() {
 
     # Cause Test name to be capitalized
     test_tag="$(sanitize_test_names ${test} | tr a-z A-Z)_TAG"
+    git_branch_tag="master"
+    if [[ "$test_tag" != *"PORTABLE"* ]]; then
+        git_branch_tag=$test_tag
+    fi
 
     # Get Github folder name
     folder="$(echo ${github_url} | awk -F'/' '{print $NF}' | sed 's/.git//g')"
@@ -394,7 +489,7 @@ print_clone_project() {
             "\nENV ${test_tag}=\$${test_tag}" \
             "\nRUN git clone ${github_url}" \
             "\nWORKDIR /${folder}/" \
-            "\nRUN git checkout \$${test_tag}" \
+            "\nRUN git checkout \$${git_branch_tag}" \
             "\nWORKDIR /" \
             "\n" >> ${file}
 }
@@ -447,7 +542,9 @@ generate_dockerfile() {
     os=$5
     package=$6
     build=$7
-    check_external_custom_test=$8
+    platform=$8
+    check_external_custom_test=$9
+
 
     if [[ ${check_external_custom_test} -eq 1 ]]; then
         tag_version=${EXTERNAL_REPO_BRANCH}
@@ -471,7 +568,7 @@ generate_dockerfile() {
     print_${os}_pkg ${file} "${!packages}";
 
     if [[ ! -z ${ant_version} ]]; then
-        print_ant_install ${file} ${ant_version};
+        print_ant_install ${file} ${ant_version} ${os};
     fi
 
     if [[ ! -z ${ant_contrib_version} ]]; then
@@ -499,7 +596,11 @@ generate_dockerfile() {
     fi
 
     if [[ ! -z ${criu_version} ]]; then
-        print_criu_install ${file} ${criu_version};
+        print_criu_install ${file} ${criu_version} ${os} ${platform};
+    fi
+
+    if [[ ! -z ${jdk_install} ]]; then
+        print_jdk_install ${file} ${os} ${platform};
     fi
     
     if [[ ! -z ${maven_version} ]]; then
